卡片系統1 && 修各種bug
明天才會有2
狗B東西真不好做

+++ Assets/UI/MobHealth.cs
@@ -8,7 +8,7 @@ public class HealthBar : MonoBehaviour
     private RectTransform rectTransform;
     private Camera mainCamera;
     public Image healthBar;
-    private int CurrentHealth;
+    private float CurrentHealth;
     private float maxHealth;
 
     void Start()
